requirements.txt
category_encoders==2.9.0
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mrmr-selection==0.2.8
numpy==2.4.2
packaging==26.0
pandas==3.0.0
patsy==1.0.2
pillow==12.1.1
polars==1.38.1
polars-runtime-32==1.38.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
six==1.17.0
statsmodels==0.14.6
threadpoolctl==3.6.0
tqdm==4.67.3
xgboost==3.2.0
